Rename 'api_server_v2' crate to 'api_server'
