+++ Cargo.toml
@@ -6,7 +6,7 @@ authors = ["Amazon firecracker team <firecracker-devel@amazon.com>"]
 [dependencies]
 clap = "=2.27.1"
 
-api_server_v2 = { path = "api_server_v2" }
+api_server = { path = "api_server" }
 devices = { path = "devices" }
 sys_util = { path = "sys_util" }
 vmm = { path = "vmm" }
